Segmentの計算を一旦外に移動
