+++ src/lib/supabaseDto.ts
@@ -1,7 +1,7 @@
 import { fetchTable, insertTable } from './supabase';
 import { DBLiftStatusView, OneDayLiftLogs, DBResort, DBLiftStatus, YukiyamaResponse, DBLift, ResortsDto, LiftsDto, liftStatus, LiftSegment, LiftSegmentsByLiftId, OperationStatus } from '@/types';
 import dayjs from '@/util/dayjs';
-import { SEGMENTS_PER_HOUR, ONE_SEGMENT_MINUTES } from './constants';
+import { ONE_SEGMENT_MINUTES } from './constants';
 import PerformanceMonitor from '@/util/performance';
 
 /* ------------------------------------------------------------
@@ -45,108 +45,109 @@ const roundMinutes = (dayjs: dayjs.Dayjs): dayjs.Dayjs => {
   return dayjs.minute(minutes).startOf('minute');
 }
 
-// リフトのログからstatus barのどの位置にstatusを表示するかを計算する（改善版）
-const getSegmentsAndGroups = (liftLogs: liftStatus[], availableHours: number[]): LiftSegment[] => {
-  if (liftLogs.length === 0 || availableHours.length === 0) {
-    return [];
-  }
+// // リフトのログからstatus barのどの位置にstatusを表示するかを計算する（改善版）
+// const getSegmentsAndGroups = (liftLogs: liftStatus[], availableHours: number[]): LiftSegment[] => {
+//   if (liftLogs.length === 0 || availableHours.length === 0) {
+//     return [];
+//   }
 
-  const now = dayjs.tz(new Date(), 'UTC');
-  const sortedHours = availableHours.sort((a, b) => a - b);
+//   const now = dayjs.tz(new Date(), 'UTC');
+//   const sortedHours = availableHours.sort((a, b) => a - b);
   
-  // 表示期間の開始時刻と終了時刻を計算
-  const baseDate = dayjs.tz(liftLogs[0].round_created_at, 'UTC').tz('Asia/Tokyo');
-  const startTime = baseDate.hour(sortedHours[0]).minute(0).startOf('minute').utc();
-  const endTime = baseDate.hour(sortedHours[sortedHours.length - 1] + 1).minute(0).startOf('minute').utc();
+//   // 表示期間の開始時刻と終了時刻を計算
+//   const baseDate = dayjs.tz(liftLogs[0].round_created_at, 'UTC').tz('Asia/Tokyo');
+//   const startTime = baseDate.hour(sortedHours[0]).minute(0).startOf('minute').utc();
+//   const endTime = baseDate.hour(sortedHours[sortedHours.length - 1] + 1).minute(0).startOf('minute').utc();
   
-  // 総セグメント数を計算
-  const totalSegments = sortedHours.length * SEGMENTS_PER_HOUR;
+//   // 総セグメント数を計算
+//   const totalSegments = sortedHours.length * SEGMENTS_PER_HOUR;
   
-  const result: LiftSegment[] = [];
+//   const result: LiftSegment[] = [];
   
-  // ログを時間順にソート（round_created_atを使用）
-  const sortedLogs = [...liftLogs].sort((a, b) => 
-    dayjs.tz(a.round_created_at, 'UTC').valueOf() - dayjs.tz(b.round_created_at, 'UTC').valueOf()
-  );
+//   // ログを時間順にソート（round_created_atを使用）
+//   const sortedLogs = [...liftLogs].sort((a, b) => 
+//     dayjs.tz(a.round_created_at, 'UTC').valueOf() - dayjs.tz(b.round_created_at, 'UTC').valueOf()
+//   );
   
-  for (let i = 0; i < sortedLogs.length; i++) {
-    const currentLog = sortedLogs[i];
-    const currentTime = dayjs.tz(currentLog.round_created_at, 'UTC');
+//   for (let i = 0; i < sortedLogs.length; i++) {
+//     const currentLog = sortedLogs[i];
+//     const currentTime = dayjs.tz(currentLog.round_created_at, 'UTC');
     
-    // 次のログの時刻（なければ終了時刻）
-    const nextTime = i < sortedLogs.length - 1 
-      ? dayjs.tz(sortedLogs[i + 1].round_created_at, 'UTC')
-      : endTime;
+//     // 次のログの時刻（なければ終了時刻）
+//     const nextTime = i < sortedLogs.length - 1 
+//       ? dayjs.tz(sortedLogs[i + 1].round_created_at, 'UTC')
+//       : endTime;
     
-    // 現在時刻より未来の場合はスキップ
-    if (currentTime.isAfter(now)) {
-      break;
-    }
+//     // 現在時刻より未来の場合はスキップ
+//     if (currentTime.isAfter(now)) {
+//       break;
+//     }
     
-    // ステータスの継続時間を分単位で計算
-    const durationMinutes = Math.min(
-      nextTime.diff(currentTime, 'minute'),
-      endTime.diff(currentTime, 'minute')
-    );
+//     // ステータスの継続時間を分単位で計算
+//     const durationMinutes = Math.min(
+//       nextTime.diff(currentTime, 'minute'),
+//       endTime.diff(currentTime, 'minute')
+//     );
     
-    // セグメント数に変換（1セグメント = ONE_SEGMENT_MINUTES分）
-    const segmentCount = Math.max(1, Math.ceil(durationMinutes / ONE_SEGMENT_MINUTES));
+//     // セグメント数に変換（1セグメント = ONE_SEGMENT_MINUTES分）
+//     const segmentCount = Math.max(1, Math.ceil(durationMinutes / ONE_SEGMENT_MINUTES));
     
-    // 現在のセグメントインデックスを計算
-    const timeFromStart = currentTime.diff(startTime, 'minute');
-    const segmentIndex = Math.floor(timeFromStart / ONE_SEGMENT_MINUTES);
+//     // 現在のセグメントインデックスを計算
+//     const timeFromStart = currentTime.diff(startTime, 'minute');
+//     const segmentIndex = Math.floor(timeFromStart / ONE_SEGMENT_MINUTES);
     
-    // 範囲内のセグメントのみ追加
-    if (segmentIndex >= 0 && segmentIndex < totalSegments) {
-      result.push({
-        status: currentLog.status,
-        created_at: currentLog.created_at,
-        round_created_at: currentLog.round_created_at,
-        startIndex: segmentIndex,
-        count: Math.min(segmentCount, totalSegments - segmentIndex)
-      });
-    }
-  }
+//     // 範囲内のセグメントのみ追加
+//     if (segmentIndex >= 0 && segmentIndex < totalSegments) {
+//       result.push({
+//         status: currentLog.status,
+//         created_at: currentLog.created_at,
+//         round_created_at: currentLog.round_created_at,
+//         startIndex: segmentIndex,
+//         count: Math.min(segmentCount, totalSegments - segmentIndex)
+//       });
+//     }
+//   }
   
-  // 時間外セグメントを埋める
-  if (result.length === 0 || result[0].startIndex > 0) {
-    // 最初のセグメントが時間外の場合
-    const outsideStatus: LiftSegment = {
-      status: 'outside-hours' as OperationStatus,
-      created_at: startTime.toISOString(),
-      round_created_at: startTime.toISOString(),
-      startIndex: 0,
-      count: result.length > 0 ? result[0].startIndex : totalSegments
-    };
-    result.unshift(outsideStatus);
-  }
+//   // 時間外セグメントを埋める
+//   if (result.length === 0 || result[0].startIndex > 0) {
+//     // 最初のセグメントが時間外の場合
+//     const outsideStatus: LiftSegment = {
+//       status: 'outside-hours' as OperationStatus,
+//       created_at: startTime.toISOString(),
+//       round_created_at: startTime.toISOString(),
+//       startIndex: 0,
+//       count: result.length > 0 ? result[0].startIndex : totalSegments
+//     };
+//     result.unshift(outsideStatus);
+//   }
   
-  // 現在時刻以降を時間外で埋める
-  const lastSegment = result[result.length - 1];
-  const lastEndIndex = lastSegment.startIndex + lastSegment.count;
-  if (lastEndIndex < totalSegments) {
-    const nowSegmentIndex = Math.floor(now.diff(startTime, 'minute') / ONE_SEGMENT_MINUTES);
-    const outsideStartIndex = Math.max(lastEndIndex, nowSegmentIndex);
+//   // 現在時刻以降を時間外で埋める
+//   const lastSegment = result[result.length - 1];
+//   const lastEndIndex = lastSegment.startIndex + lastSegment.count;
+//   if (lastEndIndex < totalSegments) {
+//     const nowSegmentIndex = Math.floor(now.diff(startTime, 'minute') / ONE_SEGMENT_MINUTES);
+//     const outsideStartIndex = Math.max(lastEndIndex, nowSegmentIndex);
     
-    if (outsideStartIndex < totalSegments) {
-      result.push({
-        status: 'outside-hours' as OperationStatus,
-        created_at: now.toISOString(),
-        round_created_at: now.toISOString(),
-        startIndex: outsideStartIndex,
-        count: totalSegments - outsideStartIndex
-      });
-    }
-  }
+//     if (outsideStartIndex < totalSegments) {
+//       result.push({
+//         status: 'outside-hours' as OperationStatus,
+//         created_at: now.toISOString(),
+//         round_created_at: now.toISOString(),
+//         startIndex: outsideStartIndex,
+//         count: totalSegments - outsideStartIndex
+//       });
+//     }
+//   }
   
-  return result;
-};
+//   return result;
+// };
 
 // LiftStatus一覧 resort_id: {yyyy-mm-dd: {lift_id: {status, created_at}}}
 export async function fetchOneDayLiftLogs(
   resortId: number, 
   currentDate: string
-): Promise<LiftSegmentsByLiftId> {
+): Promise<OneDayLiftLogs> {
+
   PerformanceMonitor.start('fetch-one-day-lift-logs');
   
   const fromDate = dayjs.tz(currentDate, 'Asia/Tokyo').toDate();
@@ -160,31 +161,33 @@ export async function fetchOneDayLiftLogs(
 
   if (!data) {
     console.error('Error fetching lift statuses: data is null');
-    return { liftSegments: {}, hours: [] };
+    return { liftLogs: {}, hours: [] };
   }
 
-  // 1. メモリ効率を改善：Mapを使用してデータを整理
-  const resortLiftLogs = new Map<number, DBLiftStatusView[]>();
+  // 1. メモリ効率を改善：オブジェクトを使用してデータを整理
+  const resortLiftLogs: { [liftId: number]: DBLiftStatusView[] } = {};
   const hours = new Set<number>();
   
-  // データを一度のループで整理（filterを削除）
+  
+  // liftIdごとにログをまとめる
   for (const log of data) {
     const hour = dayjs.tz(log.created_at, 'UTC').tz('Asia/Tokyo').hour();
     hours.add(hour);
     
-    const liftLogs = resortLiftLogs.get(log.lift_id) || [];
-    liftLogs.push(log);
-    resortLiftLogs.set(log.lift_id, liftLogs);
+    if (!resortLiftLogs[log.lift_id]) {
+      resortLiftLogs[log.lift_id] = [];
+    }
+    resortLiftLogs[log.lift_id].push(log);
   }
   
   // 2. 各リフトのログを時間順にソートし、重複除去と連続ステータス処理
-  const logsByLiftId = new Map<number, liftStatus[]>();
+  const logsByLiftId: { [liftId: number]: liftStatus[] } = {};
   
-  for (const [liftId, liftLogs] of resortLiftLogs) {
+  for (const [liftId, liftLogs] of Object.entries(resortLiftLogs)) {
     // 時間順にソート
-    liftLogs.sort((a, b) => 
-      dayjs.tz(a.created_at, 'UTC').valueOf() - dayjs.tz(b.created_at, 'UTC').valueOf()
-    );
+    // liftLogs.sort((a, b) => 
+    //   dayjs.tz(a.created_at, 'UTC').valueOf() - dayjs.tz(b.created_at, 'UTC').valueOf()
+    // );
     
     const processedLogs: liftStatus[] = [];
     let lastStatus: liftStatus | undefined;
@@ -211,25 +214,27 @@ export async function fetchOneDayLiftLogs(
         lastStatus = newStatus;
       }
     }
-    logsByLiftId.set(liftId, processedLogs);
+    logsByLiftId[Number(liftId)] = processedLogs;
   }
 
-  // 3. セグメントとグループの計算を最適化
-  const liftSegments: { [liftId: number]: LiftSegment[] } = {};
-  const sortedHours = Array.from(hours).sort((a, b) => a - b);
-  
-  for (const [liftId, liftLogs] of logsByLiftId) {
-    liftSegments[liftId] = getSegmentsAndGroups(liftLogs, sortedHours);
-  }
+  // // 3. セグメントとグループの計算を最適化
+  // const liftSegments: { [liftId: number]: LiftSegment[] } = {};
+  // const sortedHours = Array.from(hours).sort((a, b) => a - b);
+  // for (const [liftId, liftLogs] of logsByLiftId) {
+  //   liftSegments[liftId] = getSegmentsAndGroups(liftLogs, sortedHours);
+  // }
 
   const metrics = PerformanceMonitor.end('fetch-one-day-lift-logs');
   console.log('fetchOneDayLiftLogs パフォーマンス:', {
+    resortId,
     duration: metrics.duration,
-    dataSize: data.length,
-    resortDataSize: data.length
+    dataSize: data.length
   });
 
-  return { liftSegments, hours: sortedHours };
+  return { 
+    liftLogs: logsByLiftId, 
+    hours: Array.from(hours).sort((a, b) => a - b)
+  };
 }
 
 /* ------------------------------------------------------------
